+++ LivingSpace.h
@@ -11,13 +11,13 @@
 class LivingSpace {
 public:
     LivingSpace() = delete;
-    LivingSpace(float window_width, float window_length, float out_of_screen_limit = 0 );
+    LivingSpace(unsigned int window_width, unsigned int window_length, float out_of_screen_limit = 0 );
     bool operator==(const Entity& entity) const;
     bool operator!=(const Entity& entity) const;
     enum MapCorner { UpLeft, UpRight, BottomLeft, BottomRight };
 private:
-    float window_width;
-    float window_length;
+    unsigned int window_width;
+    unsigned int window_length;
     float out_of_screen_limit;
 };
